Класс App оформлен отдельными файлами
