+++ src/ctool1cd/CMakeLists.txt
@@ -4,8 +4,8 @@ project (ctool1cd)
 set (tool_VERSION_MAJOR 1)
 set (tool_VERSION_MINOR 0)
 
-set (CTOOL_SOURCES cTool_1CD_Main.cpp ParseCommandLine.cpp)
-set (CTOOL_HEADERS cTool_1CD_Main.h ParseCommandLine.h ErrorCode.h)
+set (CTOOL_SOURCES cTool_1CD_Main.cpp ParseCommandLine.cpp App.cpp)
+set (CTOOL_HEADERS cTool_1CD_Main.h ParseCommandLine.h ErrorCode.h App.h)
 
 add_executable (ctool1cd ${CTOOL_SOURCES} ${CTOOL_HEADERS})
